Renaming modules for consistency
